[chore] videoplayer 패키지 분리
