+++ app/src/main/java/com/squirtles/musicroad/videoplayer/VideoPlayerOverlay.kt
@@ -1,11 +1,5 @@
-package com.squirtles.musicroad.pick
+package com.squirtles.musicroad.videoplayer
 
-import android.graphics.Matrix
-import android.graphics.SurfaceTexture
-import android.view.Surface
-import android.view.TextureView
-import android.widget.Toast
-import androidx.annotation.OptIn
 import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
@@ -36,7 +30,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -46,7 +39,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
@@ -54,13 +46,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.content.ContextCompat.getString
 import androidx.core.graphics.toColorInt
-import androidx.media3.common.MediaItem
-import androidx.media3.common.Player
-import androidx.media3.common.util.UnstableApi
-import androidx.media3.exoplayer.ExoPlayer
 import com.squirtles.domain.model.Creator
 import com.squirtles.domain.model.LocationPoint
 import com.squirtles.domain.model.Pick
@@ -72,29 +58,9 @@ import com.squirtles.musicroad.ui.theme.White
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
-@OptIn(UnstableApi::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun MusicVideoScreen(
-    pick: Pick,
-    swipePlayState: Boolean,
-    modifier: Modifier,
-    onBackClick: () -> Unit
-) {
-    val context = LocalContext.current
-    val player = remember { ExoPlayer.Builder(context).build() }
-    val playerState = remember { mutableStateOf(true) }
-
-    Box(
-        modifier = modifier
-    ) {
-        MusicVideoPlayer(pick.musicVideoUrl, player, swipePlayState, playerState)
-        VideoPlayerOverlay(pick, playerState, onBackClick)
-    }
-}
-
-@kotlin.OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun VideoPlayerOverlay(
+fun VideoPlayerOverlay(
     pick: Pick,
     playerState: MutableState<Boolean>,
     onBackClick: () -> Unit
@@ -227,79 +193,6 @@ private fun VideoPlayerOverlay(
             }
         }
     }
-}
-
-@Composable
-private fun MusicVideoPlayer(
-    videoUri: String,
-    player: ExoPlayer,
-    swipePlayState: Boolean,
-    playerPlayState: MutableState<Boolean>,
-) {
-    val context = LocalContext.current
-    val textureView = remember { TextureView(context) }
-
-    DisposableEffect(Unit) {
-        onDispose {
-            player.release()
-        }
-    }
-
-    AndroidView(
-        factory = {
-            textureView.apply {
-                surfaceTextureListener = object : TextureView.SurfaceTextureListener {
-                    override fun onSurfaceTextureAvailable(surfaceTexture: SurfaceTexture, width: Int, height: Int) {
-                        val surface = Surface(surfaceTexture)
-                        val mediaItem = MediaItem.fromUri(videoUri)
-                        player.setVideoSurface(surface)
-                        player.setMediaItem(mediaItem)
-                        player.prepare()
-
-                        setVideoSize(width, height, textureView)
-
-                        player.addListener(object : Player.Listener {
-                            override fun onPlaybackStateChanged(state: Int) {
-                                if (state == Player.STATE_ENDED) {
-                                    Toast.makeText(context, getString(context, R.string.video_player_ended_message), Toast.LENGTH_SHORT).show()
-                                    playerPlayState.value = false
-                                    player.seekTo(0)
-                                }
-                            }
-                        })
-                    }
-
-                    override fun onSurfaceTextureSizeChanged(surfaceTexture: SurfaceTexture, width: Int, height: Int) {
-                        // TODO
-                    }
-
-                    override fun onSurfaceTextureDestroyed(surfaceTexture: SurfaceTexture): Boolean {
-                        player.setVideoSurface(null)
-                        return true
-                    }
-
-                    override fun onSurfaceTextureUpdated(surfaceTexture: SurfaceTexture) {
-                        // TODO
-                    }
-                }
-            }
-        }
-    ) {
-        if (swipePlayState && playerPlayState.value) player.play()
-        else player.pause()
-    }
-}
-
-private fun setVideoSize(width: Int, height: Int, textureView: TextureView) {
-    // 영상을 화면 크기에 맞게 확대
-    val matrix = Matrix()
-    val scaleFactor = height.toFloat() / width.toFloat()
-    matrix.setScale(scaleFactor, 1f)
-
-    // 영상 중앙 정렬
-    val translateX = (width - width * scaleFactor) / 2f
-    matrix.postTranslate(translateX, 0f)
-    textureView.setTransform(matrix)
 }
 
 @Preview
